+++ SensingKit/SKAltimeter.h
@@ -1,5 +1,5 @@
 //
-//  SKSensorModuleType.h
+//  SKAltimeter.h
 //  SensingKit
 //
 //  Copyright (c) 2014. Queen Mary University of London
@@ -22,16 +22,8 @@
 //  along with SensingKit-iOS.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-typedef NS_ENUM(NSInteger, SKSensorModuleType) {
-    Accelerometer,
-    Gyroscope,
-    Magnetometer,
-    DeviceMotion,
-    Activity,
-    Pedometer,
-    Altimeter,
-    Battery,
-    Location,
-    iBeaconProximity,
-    EddystoneProximity
-};
+#import "SKAbstractSensorModule.h"
+
+@interface SKAltimeter : SKAbstractSensorModule
+
+@end